Updated logo
* Globe adapted from https://commons.wikimedia.org/wiki/File:Globe.svg
* The original work was published under public domain:

"I, the copyright holder of this work, release this work into the public
domain. This applies worldwide. In some countries this may not be legally
possible; if so: I grant anyone the right to use this work for any purpose,
without any conditions, unless such conditions are required by law."